Refactor pulse module into clean, focused submodules
